requirements.txt
beautifulsoup4==4.8.1
bs4==0.0.1
cachetools==3.1.1
camelot-py==0.7.3
certifi==2019.9.11
chardet==3.0.4
Click==7.0
datefinder==0.7.0
et-xmlfile==1.0.1
google-api-python-client==1.7.11
google-auth==1.7.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.14.0
idna==2.8
jdcal==1.4.1
numpy==1.17.4
oauthlib==3.1.0
opencv-python==4.1.1.26
openpyxl==3.0.1
pandas==0.25.3
pdfminer.six==20191110
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycryptodome==3.9.4
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2019.3
regex==2019.11.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
sortedcontainers==2.1.0
soupsieve==1.9.5
uritemplate==3.0.0
urllib3==1.25.7
